:bug: Fixed bibliography not respected when set in dir-locals
`company-bibtex-bibliography` was not respected when set in file- or
dir-locals. Because `with-temp-buffer` would open another buffer that
would not have the locals of the current file,
`company-bibtex-bibliography` would not be present in such a case.

However, it is desirable to configure bibtex-paths in
dir-locals (e.g. on a per-project basis).

This fixes this behaviour by pulling `company-bibtex-bibliography` out
into a variable before calling `with-temp-buffer`.